+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/breve/store/actions.ts
@@ -3,6 +3,7 @@
  */
 import { askQuestionSync } from '@automattic/jetpack-ai-client';
 import { select } from '@wordpress/data';
+import { BREVE_FEATURE_NAME } from '../constants';
 import { getRequestMessages } from '../utils/get-request-messages';
 
 // ACTIONS
@@ -66,21 +67,25 @@ export function invalidateSuggestions( feature: string, blockId: string ) {
 }
 
 export function setSuggestions( {
+	anchor,
 	id,
 	feature,
 	target,
-	sentence,
+	text,
 	blockId,
 	occurrence,
 }: {
+	anchor: HTMLElement;
 	id: string;
 	feature: string;
 	target: string;
-	sentence: string;
+	text: string;
 	blockId: string;
 	occurrence: string;
 } ) {
 	return ( { dispatch } ) => {
+		anchor?.classList?.add( 'is-loading' );
+
 		dispatch( {
 			type: 'SET_SUGGESTIONS_LOADING',
 			id,
@@ -93,15 +98,17 @@ export function setSuggestions( {
 			getRequestMessages( {
 				feature,
 				target,
-				sentence,
+				text,
 				blockId,
 				occurrence,
 			} ),
 			{
-				feature: 'jetpack-ai-breve',
+				feature: BREVE_FEATURE_NAME,
 			}
 		)
 			.then( response => {
+				anchor?.classList?.remove( 'is-loading' );
+
 				try {
 					const suggestions = JSON.parse( response );
 					dispatch( {
@@ -122,6 +129,8 @@ export function setSuggestions( {
 				}
 			} )
 			.catch( () => {
+				anchor?.classList?.remove( 'is-loading' );
+
 				dispatch( {
 					type: 'SET_SUGGESTIONS_LOADING',
 					id,