+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [0.55.2] - 2024-07-26
+### Added
+- Export button props type to be used elsewhere [#38549]
+
+## [0.55.1] - 2024-07-25
+### Added
+- Added `className` prop to `Alert` component [#38450]
+
+### Changed
+- React compatibility: Ensuring createRoot is not called more than once. [#38495]
+
+## [0.55.0] - 2024-07-22
+### Removed
+- Remove compatibility with WordPress 6.4. [#38386]
+
 ## [0.54.4] - 2024-07-18
 ### Changed
 - Internal updates.
@@ -1087,6 +1102,9 @@
 ### Changed
 - Update node version requirement to 14.16.1
 
+[0.55.2]: https://github.com/Automattic/jetpack-components/compare/0.55.1...0.55.2
+[0.55.1]: https://github.com/Automattic/jetpack-components/compare/0.55.0...0.55.1
+[0.55.0]: https://github.com/Automattic/jetpack-components/compare/0.54.4...0.55.0
 [0.54.4]: https://github.com/Automattic/jetpack-components/compare/0.54.3...0.54.4
 [0.54.3]: https://github.com/Automattic/jetpack-components/compare/0.54.2...0.54.3
 [0.54.2]: https://github.com/Automattic/jetpack-components/compare/0.54.1...0.54.2